CFE-1655: process-related promises via '/proc'
Process-related promises have hitherto been implemented by invoking the
'ps' command and attempting to parse its output. But 'ps' output
is highly human-oriented, and often not easy to machine-parse reliably.
It is also highly OS-variable.

CFE-1655 had initially proposed the idea of using "/proc" where available.
My own encounter with the flaws in the existing mechanism is at CFE-3020.

This will be a gradual, evolving transition, on a per-OS "opt-in" basis,
from old ps-command working to new system-probe (os "sysprobe") working.

This first commit sets up the autotools framework to permit this functionality
to be developed but does not include the code for any OS.

To allow gradual introduction, files "libpromises/process_solaris.c" and similar
are renamed to "libpromises/process_cmd_solaris.c" to reflect their old
command-based nature.  When an OS is implemented in the new way (e.g. "/proc")
a corresponding "libpromises/process_sysprobe_solaris.c" (etc.) is introduced.

Where conditional compilation is needed, "HAVE_PS_VIA_SYSPROBE" indicates
the new way.

Although not in this intial commit, it is intended that the new system-probing
will use JSON-based data structures to separate out the OS-dependent and
OS-independent aspects; the latter could later be moved into "libntech".