Adapt the 10_files/this_promiser.cf test to better changes reporting
The outcome classes are now defined for the top-level directory
even if 'include_basedir' is 'false'. It makes more sense -- if
there are changes made in the directory, the respective
'_repaired' class is defined. And then it can be used for things
like "Do we need to update the archive/backup of that
directory?".

Ticket: ENT-5291
Changelog: The outcome classes are now defined for the top-level directory when 'include_basedir' is 'false'