sysinfo.c: Limit string sizes in sscanf calls and assert buffers
In practice, most of these calls were safe as the strings being scanned
are read from buffers in of the same size.

Still LGTM was alerting on "Unbounded write" so it is nice to fix.

Added also asserts for the sizes of the buffers.

Couple of char buffers where defined with sizes calculated at runtime;
change these to use fixed length instead, seeking consistency across the
file.

Changelog: None
Ticket: ENT-6019
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>