Make sure we don't truncate the contents of 'linkbuf' in LinkCopy()
The value from the 'linkbuf' buffer is later copied into another
buffer with the size of CF_BUFSIZE. And since there has to be
room for the terminating NUL-byte in the other buffer, at least
(CF_BUFSIZE - 1) characters can be copied from 'linkbuf' so the
contents of 'linkbuf' can potentially be truncated.

gcc warns about this.