Do aligned access to LMDB data in cf-check
Data stored in LMDB and accessed with the 'MDB_val.data' field is
a pointer to a map of the LMDB file (set up with mmap()). This
doesn't ensure proper alignment and so in order to avoid
potential SIGBUS due to bad alignment (on architectures sensitive
to these issues), we need to create local copies of the data
before trying to access them in any way than just one byte at a
time.

Ticket: CFE-3373
Changelog: None