Replace the DONTDO boolean with an eval covering multiple modes
The 'DONTDO' global variable is basically the translation of the
'--dry-run' option (some of) our binaries support. With the
'--audit' mode/option, there will be multiple possible modes
having different rules for making/skipping and reporting changes.

Let's preserve the DONTDO variable as a macro so that the
existing code using it doesn't have to be modified and so that we
can use the usage of the variable as an indicator of code that
needs to be adapted to the new modes.

Ticket: ENT-5292
Changelog: None