Do not format the message in cfPS() and Record*() functions
Logging is lazy-evaluated -- the message should only be formatted
if it is actually being logged. In many cases, the Log() function
does nothing because the log level of the message is too low to
be logged (e.g. LOG_LEVEL_DEBUG). It is thus important to avoid
the expensive work of formatting the message unless it is
actually being logged. That can be achieved by making sure that
the format and parameters are passed all the way to the logging
code with no formatting happening in advance. And that's what the
functions like cfPS(), RecordChange(), RecordWarning(),
etc. should do.