Move the functions for preparting changes chroot into a separate file
So that the PrepareChangesChroot() function can be used from
multiple different files.

Also move the required files_* files from the cf-agent directory
to libpromises. Functions in those files are no longer specific
to the cf-agent binary, they will now be required by libpromises
in general. And in general, it makes more sense to have these
files and functions in the library.

Ticket: ENT-5292
Changelog: None