Wait for background processes before generating reports
Reports generation only happens if there is exactly one
'cf-agent' process running (to avoid corruptions in the reporting
data). If some promise in the policy uses an action policy with
'background => true', there might still be a forked child process
evaluating the promise when the main process reaches the point
when it wants to generate reports. In such cases, it is important
that the main process waits for the forked child processes.

Ticket: ENT-6042
Changelog: Promises with 'action => bg()' no longer break reporting data