ENT-5335 add workarounds for tests failing on SUSE.
Issue is that `usermod` on suse doesn't like editing secondary groups together
with any other attributes. In real life, this isn't a big issue - the promise
will be completely repaired on second run. However, this test failed.

This commit changes this behavior to expect secondary group to *not* be set
after first promise run. When this get fixed, this commit should be reverted :)