Added new policy function execresult_as_data()
This function is similar to execresult, except it returns
a data container with both exit code and command output.

We have 2 ways to name functions like this:

* variablesmatching_as_data()
* data_readstringarray()

I chose the former because I think it reads better, and it also
helps with alphabetical sorting in documentation and auto complete.

We should go over all of these and make them consistent.

Ticket: CFE-3315
Changelog: Title
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>