Avoid returning an address of a local variable in GetCommonConstraint()
Assigning a struct to another local struct means copying contents
of the struct. Returning address to an array item in the local
struct looks like an invalid thing to do. In this specific case
it works because the array itself is not copied (only the pointer
to it) so the address that is returned is not inside the local
variable, but it is still confusing for developers as well as for
code checkers. Let's just use pointer to the array.

Ticket: ENT-6043
Changelog: None