Limited unqualified host and domain name to 511 characters
This is to make the truncation behavior consistent, since
fqhost is maximum 1023 characters. Thus, domain or host name
may in rare cases be truncated on their own, but fqhost will
never be truncated (compared to uqhost and domain).

Also introduced warning messages if you use host names or
domain names which are longer than 255 characters. The limit
for fqhost in DNS is 255 characters:
https://tools.ietf.org/html/rfc2181#section-11

Changelog: Title
Ticket: CFE-3409
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>