ValidateSQLTableName(): Added more error handling and error messages
When we decrease the sizes of these buffers, we really don't want
any truncation to be happening. So it's better to add errors.

Note that the only place this function is used,
it prints an error when returning false, so adding more error
messages to the places where it returns false seems OK.

Still passes unit tests.

(If we start using this function in more places, we may need to
revisit this).

Changelog: None
Ticket: CFE-3409
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>