Codecov: Disabled PR statuses
Since we have fluctuations in test coverage, codecov is
currently not useful, and in some cases creating extra noise in
PRs / diffs etc.

First attempt is to disable the status / PR interactions:
https://community.codecov.io/t/disable-all-intereactions-on-prs/1218/10

Then, we should still be able to go to codecov.io to see coverage info
but it won't create noise in PRs. If this doesn't work, I'll rip it all
out.

Changelog: None
Ticket: CFE-3410
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>