Stopped updating files promise result with WARN (notkept) when setuid files are encountered
When cf-agent encounters SETUID files that are not present in the setuid log
file ( $(sys.workdir)/cfagent.$(sys.fqhost).log ) a WARNING is emitted. Until
now the promise result was also updated in a way making the promise seem to be
not kept. This change removes the updating of the promise result so that the
real promise outcome is un-altered. The warning will still be emitted and the
file will be added to the log if it is not already there.

Ticket: ENT-6519
Changelog: Title