Python promise module: Switched to using return values
Promise outcome, and result of init / terminate etc. should be return values.
Validation Error is still an exception - it is not expected to happen,
it means there is something wrong with the policy.
Not kept, repaired and kept, can all happen during "normal" circumstances.
They are just the result of a promise, and should not be exception based.
Using return values instead of modifying the state of the object,
makes more sense and makes the code cleaner.

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>