Properly support renames with relative paths
Using 'rename => newname("some-name")' is simpler than having to
specify an absolute path. However, internally, the absolute path
is required because otherwise the result depends on CWD so we
need to construct it in the code from the original name/path.

Ticket: CFE-3537
Changelog: 'rename => newname()' now supports relative paths