Add support for handling cf-runagent run requests in cf-execd
Having the ability to trigger agent runs on specific hosts is
very useful for debugging or when some changes need to be pushed
ASAP.

With these changes, 'cf-execd' listens on a socket where
processes with proper permissions can request 'cf-runagent' runs
on given hosts. The results are returned as JSON with either the
"exit_code" and "output" keys or the "error" key.

Ticket: ENT-6182
Changelog: cf-execd now handles requests to run cf-runagent on given hosts