Travis: Disabled MacOS
Failing Mac builds have caused master to be red for some days.
This seems to be because Travis changed the installed packages
on their default image. Since we are already testing MacOS
on GitHub Actions, and they are much faster and more reliable,
I don't see a point in fixing the travis build, so I disabled it.

Had to resolve conflicts since this was already an allowed
failure on 3.15.x (but not on master).

Ticket: None
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 1539824f12107a18652c4ce8a995f5fb269b6c7f)