Ensure runagent.socket ACLs are correct in case of policy changes
cf-execd creates the runagent.socket when it starts and sets ACLs
on it based on the policy. However, when policy changes, cf-execd
is not restarted, it just loads the new policy and keeps
running. Thus it needs to detect a change in the list of allowed
users and make sure the ACLs are updated in case of a change.

Ticket: ENT-6735
Changelog: None