Add a struct for messages about packages together with their versions
In simulation modes, we need to store messages about packages
operations that will potentially be shown to the user. However,
package operations can cancel each other, e.g. when installing
and then removing the same package or installing package twice,
in two different versions, which would have different results
based on the versions. Thus we need to compare versions of the
package when deciding which messages to keep and which ones to
discard.

Ticket: CFE-3569
Changelog: None