Use a different name for the DATADIR macro
Unfortunately, on mingw builds, 'DATADIR' is a name of an enum
type. So we cannot use the same string as a name for our own
macro. Let's use CF_DATADIR as the macro name which is also what
libntech uses (changed in cfengine/libntech@ad5eeaa5969).

Ticket: ENT-6789
Changelog: None