Add 'classes' body support for custom promises
The 'classes' body can be handled on the agent side just like for
any other promise. We just need to make sure the related
attributes are not sent to the promise modules so they don't have
to deal with them in validation.

Ticket: CFE-3437
Changelog: Added 'classes' body support for custom promises