Filtered error about unknown binary data from Valgrind test for cf-check dump
Recently a measurement tracking entropy available was added to the MPF. The
measurement causes cf-check dump to emit an error when dumping the
nova_static.lmdb database which in turn causes the Valigrind check to fail.

    cf-check dump ./nova_static.lmdb
    {
        "available_entropy/proc/sys/kernel/random/entropy_avail": "
    Error: This database contains unknown binary data - use --simple to print anyway

This change filters lines including "This database contains unknown binary data"
from consideration as an error by the Valgrind check.

Ticket: ENT-6626
Changelog: None