Change SINGLE_COPY_CACHE to be a StringSet
The SINGLE_COPY_CACHE is a helper structure to remember files
that were copied by the agent because they matched a regular
expression from the SINGLE_COPY_LIST ('files_single_copy' agent
control attribute). It should not be treated as a list of regular
expressions, it should be a list of simple strings (file
paths). And because ordering doesn't matter, it can be a set
which is optimized for fast presence checking.

Also make sure the structure is not leaked by the agent run.

Ticket: CFE-3621
Changelog: files_single_copy no longer treats paths of copied files as regular expressions