Allow 'files_single_copy' body agent control attribute to be an empty list
It is a list of regular expressions identifying file that should
only be copied once. So an empty list is a perfectly valid value.

Also inform that an empty list is not a valid value for other
attributes instead of saying that the specified attribute is
invalid.

Ticket: CFE-3622
Changelog: files_single_copy body agent control attribute can now be an empty list