Log the slist from the files_single_copy body control agent attribute
So that users can see the value that is actually being used.

Also use StringEqual() instead of strcmp() and add an assert that
the value is actually and slist (enforced by the parser, but
better check twice in debug builds).

Ticket: CFE-3622
Changelog: Value of the 'files_single_copy' body control attribute is now logged in verbose logging mode