Do not create iteration variables in foreign bundles
When iterating over slists, temporary variables with the slist
reference replaced by the values from the slist are created in
the 'this.' scope. For example for the following slist:

  "list" slist => { "item1", "item2" }

and the following variable expression:

  "$(some_data[$(list)]"

the following temporary helper variables are created:

  "$(this.some_data[item1])"
  "$(this.some_data[item2])"

with the respective values, which are then used as replacements
for the variable references in the individual iterations.

That introduces a problem if the variable reference actually
contains a scope as well, e.g. "config.data[$(list)]", because
there cannot be a variable with two scopes (like
"this.config.data[item1]").

Instead of creating variables in the foreign scope (bundle), like
"config.data[item1]" and "config.data[item2]", let's mangle the
variable reference to "config___data[$(list)]" and then create
and lookup variables "this.config___data[item1]" and
"this.config___data[item1]".

Variables from the 'this' scope are automatically cleaned when
leaving the current scope (bundle, body,...).

Ticket: ENT-7029
Changelog: "source=promise_iteration" variables are no longer created in foreign bundles