Add a new 'a_entry_required' parameter to DeleteDigestFromLastSeen()
The lastseen local DB has entries of multiple types. In
particular, there are k-entries mapping hostkeys to IP addresses
and a-entries for the reverse mapping of IP addresses to
hostkeys. The ideal state is to have the entries of those two
types coherent (bijective), but it's not always possible. For
this reason, we need to be able to delete k-entries even if there
is no matching a-entry.

Do not require a-entries when deleting digests in
RemoveKeysFromLastSeen() if must_be_coherent is false.

Ticket: ENT-6490
Changelog: None
(cherry picked from commit 93b778ca3aa60e130fc5d063dee9744d398a683d)