Fixed type-limits compiler warnings
As upto is size_t its impossible for it to be less than 0.
At the same time the current algorithm doesnt allow any
underflowing to occur since it uses ref_str as an anchor.

The lowest upto can go is 0 if the bracket is found at the start,
at which point all subsequent memchr will result in NULL.

Ticket: CFE-3418
Changelog: None