Avoid double free in case of invalid header from custom promise module
The module path and interpreter are strings that are taken over
by a PromiseModule structure. However, if no such structure is
created in PromiseModule_Start(), they need to be destroyed by
the calling code. Let's just do the same if a PromiseModule
structure is created, but then destroyed inside
PromiseModule_Start().

Ticket: CFE-3651
Changelog: None