Fixed bad assertion parsing promise agent blocks
The code for parsing custom promises assumed body type is always agent.
This is not the case, and here is why:

The parser calls `ParseError` when encountering a non-agent body type.
But the parser continues parsing until a limit of 13 errors have been
encountered. Thus one cannot assume the body type is always agent.
Instead we have to handle the cases when it's not.

Ticket: ENT-7099
Changelog: None
Signed-off-by: larsewi <lars.erik.wik@northern.tech>