libpromises: Fix compilation of qdbm and tokycabinet backends
With commit e2c831609 (Add a new function for overwriting a value in
local DB, 2020-05-27), a new function `DBPrivOverwrite()` has been
introduced with implementations for the various database backends. As it
turns out, both the implementation for qdbm and for tokyocabinet are
broken and don't compile out of the box: both use a different signature
compared to the declared one, and the tokyocabinet implementation uses
wrong variable names.

Fix both issues by adapting the implementations' function signatures and
by using the correct set of variables in case of tokyocabinet.