Enabled non-object bare values inside augments variables key
When discussing this feature, we mentioned that the new
format was only a problem/incompatibility for objects.
If the value is something like a string (or list)
the old structure "just works". This also means that
in the future, when 3.18.0+ is common, we can tell users
to just rename "vars" to "variables". As long as they
don't have any objects they don't have to make further
changes. This can be added as a warning, even.

Looking at our examples it seems quite common that
augments variables are strings or lists, and quite
rare that they are objects (data containers), so for
most users, they can just rename "vars" to "variables".
(With this change).

Without this change, the code crashes (abort assertion
error) if you try to do this.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>