Add a verbose log message if socket creation fails in cf-serverd
It's only an error if all socket creations or binding attempts
fail and there already is a log message for this case (see
cfengine/core@1306ec2e856e). But for debugging purposes or for
customers to be able to see what's going on, we should have more
detailed logging.

Ticket: ENT-7362
Changelog: None