Changed log level of message about reaching maximum recursion from ERROR to VERBOSE
If I promise to copy a directory recursively up to a depth of N and there are
levels deeper than N the promise should not error. Not promising enough depth to
reach the bottom isn't an error, at most it's a message more relevant for
verbose logs, if not debug logs.

Ticket: CFE-3723
Changelog: Title