Use JSON_NOT_NULL()/NULL_JSON() to better handle augments/CMDB data
JSON has a special value 'null' which represents and empty
value. We should thus treat it as such -- in most cases this
means the same as if the child object was not present. For
example with this JSON::

  "obj": {
    "child" : null
  }

the 'child' sub-object should be treated as a potential 'child2'
sub-object which is not present at all. That is, when handling
JSON/CMDB data.

Ticket: ENT-7434
Changelog: 'null' JSON value is now handled as empty data in augments/host-specific data
(cherry picked from commit cfafd2c7a504cb0234f9317394ba5bbe197d505f)