Avoid excessive logging from in-memory changes in files promises
File editing happens in two phases -- the changes are modeled in
a memory buffer and then the buffer is compared with the file on
disk to see if any changes need to be made. If the buffer and the
file differ, the buffer is committed to disk (as the file).

Thus, in places where the changes in the memory buffer are done,
it's not yet clear if they should be shown to the user as 'info:'
messages and so they need to be buffered and only actually logged
if some changes are committed to disk.

Ticket: CFE-3708
Changelog: Only real changes in files now produce info messages
(cherry picked from commit 1b647c2c32735992c392b680f4597bd7b1005676)

Conflicts:
  libntech