Add a special SELinux type and domain for Mission Portal action scripts
The notification scripts that can be hooked up to alerts are
supposed to be able to do basically anything on the system. Thus
we need to allow that, but we shouldn't enable our httpd and PHP
do anything on the system.

Ticket: ENT-8029
Changelog: None
(cherry picked from commit ef652c0d21c04fc5792e34075a52849d76497010)