Added bootstrap/failsafe exclusion for directories named .no-distrib
With build.cfengine.com and cfbs released, we would like to be able to
distribute compliance reports as cfbs modules. We don't need (or necessarily
want) to distribute compliance report definitions to every host since they are
only useful for an enterprise hub, so we introduced this directory '.no-distrib'
as an excluded directory for bootstrap and failsafe.

With this change bootstrap/failsafe will exclude any directory named .no-distrib.
Note, until the update policy is updated with the same
exclusion, any policy update will receive the files in those directories.

Ticket: ENT-8079
Changelog: Title