Include attributes of type data in PromiseRuntimeHash()
Apart from string and slist attributes (constraints), there are
also container attributes, i.e. attributes of type data. These
should generally be included in calculation of a promise hash
like any other attributes. They are stored as JSON so we can just
add the string representation of the JSON to the data sent to the
hashing function.

Currently, the only standard attribute of type data is
template_data which should actually be excluded (because it
defaults to datastate which contains volatile values like
$(sys.date) etc.), but custom promises typically use container
attributes and custom bodies are converted to such attributes
too.

Ticket: CFE-3434
Changelog: None