Added StringIsBoolean()
So, there is an issue with BooleanFromString():
* It defaults to true if you give it something unexpected,
  like an unresolved variable string.
* It is used in some places without checking first if the
  argument is a valid boolean or not.

This causes an issue where ignore_missing_inputs and
ignore_missing_bundles will default to true for
unresolved variables, but they are supposed to default
to false.

To address this, I am introducing a new function to be
used together with BooleanFromString(), to check whether
something is actually boolean.

I tried doing this without allocations and also
to account for potential changes in CF_BOOL.
But after doing it and addressing the edge cases,
I realize how hairy this is, and think that perhaps
there is a better way. See more refactoring and
fixes in subsequent commits.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>