Added config to EvalContext
Since this was already in a comment that somebody needed it
in the past, and I need it now, I went ahead and added it.

Had to adjust diff slightly because of changes in surrounding
code (restrict keys functions).

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 6cb801bb34b526c35031a3451c1598a391f10004)