Fallback to running 'getent' if user/group info lookup fails
The functions getpwnam(), getpwuid() and similar for groups may
provide incomplete results when used from our processes due to
conflicting versions of libraries that we bundle and that the NSS
plugins require. So we need a child process without our libraries
loaded to do the query. However, a program we can use for that
already exists and it's called 'getent'. Let's use it for the
fallback mechanism.

Ticket: CFE-3937
Changelog: CFEngine processes now try to use getent if the
           builtin user/group info lookup fails
(cherry picked from commit 4f1288aeae336628593c3657c2ffe5208f587e24)