Removed excessive check for content
Removed a check for whether or not the content attribute is NULL in
WriteContentFromString function in verify_files.c. This check was
excessive, due to the fact that the content attribute not beeing NULL is
the condition in which this function is called. Added an assert instead
to make this explicit.

Ticket: None
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>