Add more details to the warning in PromiseGetConstraintAsBooleanWithDefault()
When a default value for an attribute that should be set
explicitly is used, a warning is emitted. It should be easier for
the user to identify which promise exactly it is coming from.

Ticket: CFE-951
Changelog: None
(cherry picked from commit 4025aa1ae1fbeada324c67e95ae929c928a5a6a1)