Added test for ENT-8788
This change contains 2 tests.
One (tests/acceptance/31_tickets/ENT-8788/1/main.cf) illustrates the issue as
described in the ticket. It is marked as a soft-fail since it is not working.

The other (tests/acceptance/31_tickets/ENT-8788/2/main.cf) illustrates the
workaround described in the ticket. It is currently passing and is not marked to
soft fail.

Ticket: ENT-8788
Changelog: None
(cherry picked from commit 465f41d02f453bcdb4e213827d3fef4a3bf52736)