Ignore namespaces when looking up special-scoped variables
Otherwise iteration variables (values of slist/array items) that
exist in the 'this.' scope fail to be found because they ignore
namespaces. And so without this change, things like this fail:

  body file control
  {
    namespace => "test";
  }

  bundle agent namespace_check
  {
    classes:
      "server" expression => "any";

    vars:
        "packages" slist =>  { "mysql-client", "mysql-common" }, policy => "free";
        "server_packages" slist => { "mysql-server", "mysql-server-core" }, policy => "free";

        "all_packages" slist => { @(packages), @(server_packages) }, policy => "free";

    reports:
        "$(all_packages)";
  }

"$(all_packages)" is looked up as "$(test:this.all_packages)" but
the values from the slist are associated with the
"$(NULL:this.all_packages)" variable reference in the individual
iterations.

Similar thing applies to 'edit.' variables and generally
special-scoped variables except for 'def.' which is actually not
so special.

Ticket: ENT-8817
Changelog: None