Avoid excessive repeated calls of strlen() on persistent strings
When working with a string one character at a time to only do one
iteration over it, using `strlen(s)` in the for-loop's condition
really ruins the game because every single iteration of the loop
actually iterates over the whole string.

Similarly, in many places, let's just avoid calling `strlen(s)`
over and over on a string that is not modified by the surrounding
code.