Adjust cf-check and package module code for empty updates list
In some cases the response from the module results in empty
data and sometimes no data at all. Not sure why.

Regardless it makes sense to refactor the package module handling
code to save the updates <inventory> key with a zero length value
in this case instead of sometimes a zero length and sometimes a
`\n` (newline) character.

Adjusted cf-check to allow zero length values in both dump and
validate portions of the code.

Ticket: ENT-9050
Changelog: title