Add support for action_policy attribute in custom type promises
As long as the custom promise module advertises the 'action_policy'
feature by sending the appropriate feature flag in the initial
exchange header, promises of the custom types the module implements
can use the 'action_policy' attribute. Similarly, promises of its
custom promise types can be used in the --dry-run/--simulate modes. If
the module doesn't support the features, promises of its promise types
should fail validation and not be evaluated because they might not
respect the attribute or evaluation mode and make changes to the
system.

Ticket: CFE-3433
Changelog: Custom promise modules can now support the 'action_policy'
           feature allowing promises of their custom types to be used
           in dry-run and simulation modes and in combination with
           'action_policy => "warn"'.