Changed bootstrap policy to preserve log level for debug, verbose, and info
With this change if the agent bootstraps with debug, verbose, or info level
logging, sub-agent executions request the same log level. This can help
facilitate debugging of bootstrap issues.

Note: This change applies to failsafe as well, but as failsafe is an embedded
call, it wont currently happen unless a separate change tries to align failsafe
execution with the agent execution that failed.

Ticket: CFE-4121
Changelog: Title