Made cf-support use coredumpctl for core analysis only when configured in kerenl.core_pattern
It's possible for coredumpctl to be present yet the system not correctly
configured to use systemd-coredump which results in coredumpctl not being able
to find core files. This change restricts the use of coredumpctl for when
systemd-coredump is present in sysctl kernel.core_pattern.

Ticket: ENT-9985
Changelog: Title
(cherry picked from commit 9f538452fa3f361095356deabfa3911f59211e0a)