Variables & Classes Modules automatically tagged
According to the docs; all variables and classes will be tagged with
`source=module`. This was apparently only the case when specifying tags
using `^meta=Tag1,Tag2`. Now this tag is added all the time, even when
you don't specify tags.

Furthermore, we added an additional meta tag
`derived_from=<SOME_COMMAND>` which holds the path to the module script
when using the `usemodule()` policy function or the command when using a
commands promise with `module => "true"`.

Ticket: ENT-7725
Changelog: Body
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>