Added ssh_exec_t to selinux policy
Added with cfbs type+domain+label for cfbs in commit ef26390
but we didn't cherry-pick that back to 3.18.x because cfbs is not present there.

A later commit: 3695a8a, Allow our policy-driven daemons to getattr() /usr/bin/ssh
depends on ssh_exec_t so we must add it in.

Ticket: ENT-10258
Changelog: none