package[update]smatching now looks for software inventory databases
If the default software inventory from context is not specified, the
package[update]smatching policy functions now looks for the software
inventory databases in the state directory and uses them if found. This
change enables the usage of these functions in standalone policy files
without the demand for specifying the default package inventory
attribute in body common control. However, you still need the default
package inventory attribute specified in the policy framework for the
software inventory databases to exist in the first place and to be
maintained.

Ticket: ENT-9083
Changelog: commit
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>