Add a new small compatibility library
With GPL-licensed code taken from libntech/libcompat.

This also means that we have to leave all the function-replacing
macros to libntech/configure, but that's the correct thing to do
because libcompat is part of libntech and so libntech/configure
actually determines which fuctions/code will be part of
libcompat. The only exception are the *printf() functions because
their replacements are handled by special macros (see
m4/snprintf.m4 for gore details) and we need those bits to happen
in this repo too.