Added version_compare() policy function
This function can be used to compare 2 arbitrary semver version
numbers. This can be useful if you have 2 versions of a package
and you want to know if they are the same version, if one is
newer than the other etc.

Internally, it uses the same version comparison logic as
the cf_version functions (cf_version_minimum, etc.). This means
that only the major.minor.patch parts of the version string
are compared, the rest is discarded, so
3.2.1 = 3.2.1-1 = 3.2.1-2 etc.

Depends on:
https://github.com/NorthernTechHQ/libntech/pull/196

Ticket: CFE-3991
Changelog: Title
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>