Fixed infinite loop on error bug while reading interface exception file
Fixed bug where the agent may enter an infinite loop while trying to
read the interface exception file. The problem comes from the fact that
the while loop for reading each line in the file checks for end-of-file,
but not error. If an error occurs, end-of-file will never be reached,
hence the loop will never break.

Ticket: None
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>