Only kill potential CFEngine lock holders
If a non CFEngine process has a matching PID and start time, we
shouldn't try to kill it because it's practically impossible that
it is a real holder of one of our locks in cf_lock.lmdb. Most
likely it's an unfortunate process that ended up with the
matching PID and start time after a reboot.

Ticket: CFE-3982
Changelog: Only CFEngine processes are now killed as expired lock owners

Co-authored-by: Benoît Peccatte <benoit.peccatte@normation.com>