Fixed inventoried policy release id when masterfiles-stage.sh deploys with cfbs
When policy is tagged (with cf-promises -T) cf-promises looks for .git/HEAD and
uses it's content as the policy release id in the generated
cf_promises_release_id file. If .git/HEAD does not exist the policy tree is
hashed and that value is used instead.

For traditional deployments straight from version control, this works. When
deploying from cfbs .git/HEAD needs to be copied to out/masterfiles or it won't
be part of the tagged policy. This change adjusts the target appropriately when
deploying with cfbs.

Ticket: ENT-10832
Changelog: Title
(cherry picked from commit 89ab485968e1eb420c9659395cb2f334a60935f7)