Handle MDB_MAP_FULL
If an LMDB database is full, we get a special error from LMDB
when trying to use the DB giving us an option to handle the
situation.

We could try to prune/purge the DB, but there's no general rule
for which data to preserve so we just move the file aside with a
`.rotated.$timestamp` extension.

Continuing (restarting, really) with a new empty LMDB file is
better than getting stuck. So it's a last-resort solution, but
our LMDB files should never get full.

Ticket: ENT-8201
Changelog: Full LMDB files are now handled gracefully by moving
           them aside and using new empty LMDB files