Make sure the dbid enum starts at 0
This is most likely already the case on all the platforms we
support, but making it guaranteed is better. This allows us to
actually rely on the fact e.g. when we want to identify the DB by
ID from Python or a similar environment that doesn't have access
to the enum.

Also assign values to all items in the enum so that it's easier
to do the conversion visually. Plus reformat the code to make it
nicer.