Added container chroot agent materials to contrib
Working towards providing visibility in and around containers this provides a way to bring up a container which mounts the hosting root filesystem and then installs CFEngine and runs it inside of chroot in that container.

It should not affect the host other than creating an empty /var/cfengine directory.

Ticket: ENT-11345
Changelog: none