Adapt files_copy_test.c to efficient data/file copying in libntech
Copying should respect holes in sparse files, but if the original
file is not sparse, it's not a goal of the copy code to create a
sparse copy. So to test respecting holes in sparse files during
copying, we need to create the source files as sparse files.

Ticket: CFE-4380
Changelog: None