Add a built-in timeout mechanism to valgrind-checks
If things go wrong and a run of valgrind-checks gets stuck, it's
hard to terminate and clean after it properly because it runs as
root. In a container, but that doesn't make a difference. To
avoid such issues, let's make sure the `bash` process running the
valgrind checks kills itself after a period of time unless it
finishes properly (succeeding or failing).

Ticket: CFE-4416
Changelog: None