Added acknowledged field to lastseen DB
This field can be used to acknowledge an update to an incoming or
outgoing host entry in the lastseen DB. The field is set to `false` when
a key-value pair is updated. `cf-hub` may at any time acknowledge that
the update has been observed by setting the field to `true`.

Ticket: ENT-11838
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>
Co-authored-by: Craig Comstock <craig.comstock@northern.tech>