Removed check for filename too long
There was a check to see if the destination filename was larger than
`CF_BUFSIZE - 20` when copying files over the network. This makes no
sense to me.

At first glans, one might think that it has to do with the network
protocol. I.e., `GET $SIZE $FILENAME` has to fit in `CF_BUFSIZE`.
However, this is the source filename, and it is never sent over the
protocol, as far as I can see.

Tried to use git blame to see why it was introduced, but that was not of
much help. It appears to date back to the very first implmentation of
the files promise in 2008. I'll try removing it.

Ticket: None
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>