Removed libstr from tests/unit/Makefile.am
The library was probably made as a convenience library. However, when
linking with `libstr` and other CFEngine libraries, or when including
sources from these libraries, we can quickly run into many problems.
Problems like, objects 'created with both libtool and without'. Or
duplicate symbol definitions. It's better to either include with just
the sources we need, or to only link with the CFEngine libraries when
creating test binaries.

Ticket: CFE-4471
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>