Added option to choose protocol version in cf-net
Added option to choose/force the CFEngine protocol version to be used in
cf-net.

Currently, there is no way to test older protocol versions without
targeting an older server. This option will make that possible.
Furthermore, it can be used as a workaround in potential future cases
where we issue a new protocol that turns out to not work.

Ticket: ENT-12519
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>